Extract recording handling to separate class
